+++ src/ngp.cpp
@@ -17,6 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
 #include <fx.h>
+#include <stdint.h>
 #include "CommandLineHandler.h"
 #include "ComboBox.h"
 #include "FileDialog.h"
@@ -259,7 +260,7 @@ NGPFrame::onShowInfo(FXObject *, FXSelector, void *)
 long
 NGPFrame::onOverride(FXObject *, FXSelector, void *a_data)
 {
-  if ((int) a_data)
+  if ((intptr_t) a_data)
     {
       m_overrideNGP->enable();
       m_overrideHeaderSize->enable();
